update: 롤백: petservice put 이미지 로직 수정

+++ src/main/java/com/example/withdogandcat/domain/pet/PetService.java
@@ -69,10 +69,12 @@ public class PetService {
             throw new BaseException(BaseResponseStatus.USER_NOT_FOUND);
         }
 
-        imageS3Service.deleteImages(pet.getImages());
-        pet.clearImages();
-        List<Image> newImages = imageS3Service.uploadMultipleImagesForPet(imageFiles, pet);
-        newImages.forEach(pet::addImage);
+        if (imageFiles != null && !imageFiles.isEmpty() && !imageFiles.stream().allMatch(MultipartFile::isEmpty)) {
+            imageS3Service.deleteImages(pet.getImages());
+            pet.clearImages();
+            List<Image> newImages = imageS3Service.uploadMultipleImagesForPet(imageFiles, pet);
+            newImages.forEach(pet::addImage);
+        }
 
         pet.updatePetDetails(
                 petRequestDto.getPetName(),
